yazshop/requirements.txt
amqp==5.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
blessed==1.21.0
celery==5.3.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==45.0.6
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.9
Django==4.2.7
django-allauth==0.57.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==25.1
django-import-export==3.3.4
django-js-asset==3.1.2
django-picklefield==3.3
django-q2==1.8.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
filelock==3.18.0
gunicorn==21.2.0
idna==3.10
jinxed==1.3.0
kombu==5.5.4
MarkupPy==1.18
oauthlib==3.3.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
Pillow==10.0.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.7
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
PyYAML==6.0.2
redis==4.6.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tablib==3.5.0
types-python-dateutil==2.9.0.20250516
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
whitenoise==6.6.0
xlrd==2.0.2
xlwt==1.3.0
reportlab==4.2.5
dj-database-url==2.1.0
